fix(cli): quote --jobs error token with Go %q semantics like pflag/strconv (review: jobs %q thread)
