+++ apps/cli/src/legacy/commands/storage/storage.errors.ts
@@ -1,6 +1,7 @@
 import { Data } from "effect";
 
 import { legacyAqua } from "../../shared/legacy-colors.ts";
+import { legacyGoQuote } from "../../shared/legacy-go-quote.ts";
 
 /**
  * Domain errors for `supabase storage ls/cp/mv/rm`, mirroring the Go error paths
@@ -57,13 +58,19 @@ export class LegacyStorageUnsupportedOperationError extends Data.TaggedError(
  * Byte-matches pflag's `invalid argument %q for %q flag: %v` template with
  * the shorthand-prefixed flag name (`pflag/errors.go:108-116`), carrying the
  * RAW token (so `--jobs=-01` reports `"-01"`, not a normalized `"-1"`) and
- * strconv's cause (`invalid syntax` / `value out of range`). Thrown from the
- * flag's own `Flag.mapTryCatch` in `cp.command.ts` so the rejection happens
- * during command parsing, like Go's — `formatInvalidValueMessage` surfaces
- * the resulting `CliError.InvalidValue`'s message verbatim.
+ * strconv's cause (`invalid syntax` / `value out of range`). Both token
+ * occurrences are `%q`-quoted like Go's — pflag applies `%q` to the value
+ * and strconv's `NumError.Error()` wraps `e.Num` in `strconv.Quote`
+ * (`strconv/number.go:258-260`) — so an escapable token stays one escaped
+ * line (go1.26: `--jobs 'a"b'` → `… "a\"b" …`, not a raw quote/newline).
+ * Thrown from the flag's own `Flag.mapTryCatch` in `cp.command.ts` so the
+ * rejection happens during command parsing, like Go's —
+ * `formatInvalidValueMessage` surfaces the resulting
+ * `CliError.InvalidValue`'s message verbatim.
  */
 export function legacyStorageInvalidJobsMessage(token: string, cause: string): string {
-  return `invalid argument "${token}" for "-j, --jobs" flag: strconv.ParseUint: parsing "${token}": ${cause}`;
+  const quoted = legacyGoQuote(new TextEncoder().encode(token));
+  return `invalid argument ${quoted} for "-j, --jobs" flag: strconv.ParseUint: parsing ${quoted}: ${cause}`;
 }
 
 /** `cp`'s remote→remote branch (`internal/storage/cp/cp.go:57`). */
